Remove unnecessary exception guards
We don't need to catch the unspecified-password-store scenario because
it turns out to be unexpectedly useful. This enables the user to see the
contents of all of their password stores at once. If the user tries to
initialize a password store without specifying a directory, the script
will fail with an error message, so this scenario is also already
covered.

We also don't need to catch the directory-not-found scenario because we
want to be able to initialize the new password store without having to
create the dir first. This is handled by `pass` if we let it be, and it
makes for a less tedious UX.